[#515] Add v70000/gtm9370 test to verify div-by-zero errors pushed to runtime
Note - when merging V7.0-000 into YottaDB, the entire update for GTM-9370 was thrown out as YottaDB
had already fixed the issue (differently) in YDB#828 (a clearing house for issues detected by
fuzz-testing) in commit YDB@f2f3c0b21 that went into r1.34. This test already existed by the time
this was discovered. Since the new test is marginally better than the one initially written for this
in that it uses random expressions, we are keeping this test too.

The purpose of this test is to verify that the compiler does not deal with literal evaluations that
result in a divide-by-zero and rather pushes them into the runtime. To test that, we write each expr
to a separate file, then compile each of them (expecting NO error), then we run each of them and
expect an error - except we do not always get errors since some short-circuiting of boolean exprs
still occurs even with gtm_boolean and/or gtm_side_effects so while we expect errors when we run the
compiled expression, we do not require it. So long as ONE of the expressions failed when the
expressions are executed, we consider the run a PASS. This is because EVERY expression is expected
to fail when it runs as it has a divide-by-zero term in it. But because of the nature of random
expression generation, it is possible that div0 term gets put in some place where normal conditional
expression short-cutting happens so it is possible that the expression evaluation never hits the
div0 term. In that case, the expression would not fail. We allow this but require at least ONE of the
expressions to fail at runtime.

Currently, we only generate 10 expressions. Usually, with the --boolean and --side_effects=1 settings
both set, we don't get that many expressions where the divide-by-zero component is in an unreachable
part of the expression (due to short-circuited boolean evaluation) but if, on some systems, there is
a pathology that creates more than usual of these types of expressions and the test is failing,
increasing the number of statements generated by changing the maxExprCnt variable at the top of
gtm9370 from 10 to something larger can mitigate the issue.

This issue required some updates to com/BooleanExprGen.m. The following features were added:
- The Init^BooleanExprGen entry point gets an option parameter containing generation flags. In this
  instance, the flag we support is "DIV0" so we have an emphasis on generating divide-by-zero type
  expressions. We require at least one div0 type reference in every generated expression. Note that
  the Init routine had a lot of churn as some things needed to be reordered because we needed the
  basics initialized before initializing the rest.
- The CreateRoutine^BooleanExprGen entry point gets an additional optional parameter that says who
  is creating the routine. This information replaces the BooleanExprGen name in the created-by
  output message of the generated program if it is supplied.