[#514] New v70000/gtm9368 test verifies MUPIP REPLIC -SOURCE -SHUTDOWN can be stopped by Ctrl-C
This test initiates replication, does a few sets then starts up the following command in the background:

     MUPIP REPLIC -SOURCE -SHUTDOWN -TIMEOUT=95 [-ZEROBACKLOG]

Note, this invocation is done in a loop where the first time we do not specify `-ZEROBACKLOG` and
the second time we do.

Throughout the entire test, we have imptp running to cause a bunch of dirty buffers that will allow
the second iteration of the loop with the -zerobacklog option specified to be meaningfully tested.