[#511] [#512] [GTM-9358] [GTM9361] [YDB#780] Add new test v70000/gtm9358andgtm9361 to verify MUPIP REPLIC -SOURCE -SHUTDOWN changes
There were some issues trying to write this test. Originally wrote it as a non-MSR test but
couldn't set the heartbeat interval with the -CONNECTPARAMS option as none of the test system
tools supported it. Also, no tools supported specifying a non-zero timeout for source server
shutdowns or allow specification of -ZEROBACKLOG. Turns out also the MSR control scripts to
set up replication do not support those abilities either.

So this test uses brute-force methods, creating our own global directories, subdirs,
databases, etc running each command needed to create a replication environment manually
(i.e. not using any scripts from com/). Once the replication environment is created,
we do a few updates and wait for them to flush, then shutdown the source server with
the -ZEROBACKLOG option as well as -TIMEOUT=10. Since we set -CONNECTPARAMS heartbeat
value to 2 (from 15), we should be able to recognize a zero backlog within 2 seconds.

GTM-9358 - This portion of the test is composed of:
1. Finding the up/down jnlpool semaphore.
2. Verifying it is cleaned up after the source server is shutdown.
3. Verifying the shutdown log contains the new "Initiating ZEROBACKLOG shutdown operation"
   message.

GTM-9361 - The portion of the test is composed of:
1. Verify the shutdown for GTM-9358 also contained the REPL0BACKLOG message (there was no backlog)
2. Verify a new shutdown for the source server with a known backlog contains the REPLBACKLOG
   message.
3. Verify if do shutdown with TIMEOUT=3601 generates an INVSHUTDOWN error.
4. Verify if do shutdown with TIMEOUT=3600 this succeeds.
5. Verify if do shutdown with default timeout, it waits for 120 seconds.

Note, also made minor updates to gtm9302.csh and gtm9368.csh to change their `grep` invocation
to `$grep` instead.