[YDB#780] [#504] Add new v70000/gtm9340 test to verify VIEW/$VIEW() truncate > 31 char region names and gives VIEWREGLIST error if multiple regions given to $VIEW()
This test creates a region with a 31 char name (max region length) as GDE won't allow
creation of anything larger. In the M code that references this region via VIEW and $VIEW() this large name
region is specified with a > 31 char name. The test verifies this reference is truncated to 31 rather than
giving an error that the region does not exist. This is inline with how M treats variable names.

The test specifies TWO region names to the VIEW command with the POOLLIMIT option and verifies that this is
accepted. The same thing is then tried with the $VIEW() function and we verify we get the VIEWREGLIST error
and then finally verify that specifying a single too-long region name to $VIEW() succeeds. We also verify
that running VIEW and $VIEW with a too-long region name that does not exist does the proper thing (NOREGION
error).

Test run 100+ times to verify it.