[#528] New sudo/olderversion test for testing that ydbinstall can successfully install older versions
This test is designed to catch regressions in ydbinstall when ydbinstall
is actually used to download an install from gitlab (which is the most
common use case). All of our sudo tests test ydbinstall functionality when
ydbinstall is part of the YottaDB tarball; in that case, the versions of
everything it uses are all bundled with YottaDB.

This test is very minimal right now. It prints `$zroutines`, `$zchset`,
and `$zyrelease`. This would catch the issue now fixed in MR
https://gitlab.com/YottaDB/DB/YDB/-/merge_requests/1360.