Fix v70000/gtm9358andgtm9361 test with regards to removing timestamps from replication logs
Issue came up when the current date of the test went from 'July  9' to 'July 10' (note two
spaces before '9' in the first date so the date is always the same length. This test was
using a cut command with fields being specified and a delimiter as a space. When the date
changed to a two digit day, this threw off the field number calculation causing the cut
command to truncate an extra token off the front of each line. This caused the test to
fail. All cut commands in this routine have been changed to use a character position
instead.