Fix v70000/gtm6952 test to extract correct -buffsize= values from server start commands in replication log files
Like in another test, the cut command used to locate the -buffersize= option specified on replication
server start commands (both source and receiver servers) was incorrect. The cut command does not count
adjacent white space characters as one delimiter so when the timestamp at the start of the log record
had the day of month change from 9 to 10 (or more precisely from ' 9' to '10', it caused a change in
the field number to extract the -buffersize option. This caused the cut command to trim an extra token
off the the front of the line which caused the test to fail. This test now uses a simple awk invocation
to print the 12th token but awk considers adjacent white space to be a single token so it works more
consistently.