[#528] Disable sudo/olderversion on RHEL 7.9, OpenSuse Tumbleweed, Ubuntu 20.04
`sudo/olderversion` uses `ydbinstall` to download releases that are stored
on Gitlab; `ydbinstall` does that for releases that are supported ONLY;
it will complain when run on other systems as unsupported.

We use existing variables to disable on RHEL 7.

On OpenSuse Tumbleweed, we needed to enhance the test system with the
variable `gtm_test_linux_suse_distro` and create an AWK marker for
outref files called `SUSE_TUMBLEWEED`.

On Ubntu 20.04, we re-used the now unused AWK marker for outref files
`UBUNTU_21.10`, which is now renamed `UBUNTU_20.04`.

`UBUNTU_21.10` was unused in test r130/ydb534, and was removed from
there.