Restrict v70000/ydbtest350 to DBG builds only as the BLOCK and IMAGE parms of MUPIP/DSE only available in DEBUG
This test uses the MUPIP JOURNAL -EXTRACT command with the -BLOCK option to fetch an extract of a given
DB block out of the journal file. This option is only available with a debug build.

This test uses the DSE DUMP command with the -IMAGE option to format the extracted block above into
normal DSE DUMP format but the -IMAGE option is only available in a debug build.

So this test was restricted to debug builds and bypasses the PRO runs.