Fix v70000/gtm9368 to no longer use /usr/bin/kill but use (kill -INT $shutpid) to isolate kill command
When the kill command is used in a tcsh script, if the process being killed does not exist, the kill
command (since it is an internal tcsh command) kills the script. We needed gtm9368 to NOT fail at that
point as it needed to shutdown replication and running imptp processes. We initially tried changing the
command to /usr/bin/kill but that did not work on an Ubuntu 20.04 system where the kill command was
located in /bin/kill instead. Instead of worrying about that difference, this change puts the KILL
command in parens so it forces it into a subshell that then terminates but we can check the exit
code and continue in the main script.