[#529] [#501] Modify manually_start/gds_blk_max test to use V6 DBs for LOWSPC and V7 DBs for max-blk check
This existing test was modified slightly so it still runs V6 mode blocks for TEST CASE 1. We run this test
of the LOWSPC messages occurring properly on a V6 mode DB instead of a V7 mode because doing this test in
a V7 database where the database can be over 17 times what it was in V6 means it must use the $ydb_test_4g_db_blks
envvar to create a hole in the DB. However, when this facility is used, due to the way the hole is implemented
(where the local bit maps say the DB is full but the free blocks counter shows the blocks in the 'hole' as free
instead of allocated, it keeps the implementation somewhat simple but it means the LOWSPC and DBFILEXT messages
are not output. So we can't test those messages with V7 without doing a lot of work to the DB 'hole' logic that's
not really worth doing.

For the second test case though, where we attempt to extend 1 more block and expect to error out due to there
not being any more space, we create a sparse V7 database and fill it and THEN try to extend one more block.