[YDB#88] Fix various expect scripts to not rely on gtm/mumps aliases; Fix few expect race conditions
Various expect tests needed to have expect/tcsh issues fixed. This has nothing to do
with readline, but needed in order to get the tests to pass on my
machine. If you don't do this, you rely on the user's settings in order to start
a test; they did not run successfully on my machine. The list of changed
tests are:

   - expect/D9F05002548 (-f/expect race conditions)
   - expect/erasebs (-f/invalid executable/^? is a TCSH shortcut and
     needs to be escaped)
   - expect/terminal (-f)
   - v53004/ctrlc (-f, invalid executable)
   - v62000/gtm8003 (-f, invalid executable)
   - v63000a/gtm8454 (-f, invalid executable)