[YDB#88] Changes to support readline in YottaDB
This commit contains a lot of changes to support testing YottaDB with
readline integration.

Read YDB#88 first to get a background of how it works.

There are some global changes to the test system:

- Unset `ydb_readline` at start of testing to avoid having test env
  polluted by a user's env.
- Set random setting (50/50) for `ydb_readline` if not specified with
  `-env` flag. If you want to specifically test with `ydb_readline` set
  to a specific value, you must pass `-env ydb_readline=0/1`.
- Set `INPUTRC` to `/dev/null` if `ydb_readline` is on, to avoid having
  tests fail due to a user's `~/.inputrc`, which customizes how Readline
  behaves.
- Set `gtm_tst_readline_version` to current readline version using
  `ldconfig -p` and some sed magic (used to disable tests r132/ydb704
  and V52000/D9G12002636).
- Create AWK defines YDB_READLINE, YDB_READLINE7_OR_LESS,
  YDB_READLINE6_OR_LESS.
- In the email message for success or failures, add `RL` mnemonic for
  ydb_readline.

The next list is a list of changes for specific tests.

1. Update ref files dse/dse_crit and v53003/D9I10002703 for readline code changes

dse/dse_crit updated for this:

```
cat >! dse_ip  << CAT_EOF

crit (there are two DSE before this)
crit -rele (extra DSE> Before this line is executed)--only one in my
version
q

CAT_EOF

dse << DSE_EOF

crit -seize
crit
spawn "$gtm_exe/dse < dse_ip" <<<-- One less "DSE> " prompt
crit
crit -rele
crit

DSE_EOF
```

I tried to find the reason for the one less DSE prompt,
but I couldn't really figure it out, and I thought it is not worth
spending time on that, as it's a remote case and not reproducible
outside of the test system; making me think it's a race issue.

2. v53003/D9I10002703 updated to add `ydb_readline` env var.

3. Fix r120/readcmdrecallhist (multiple issues).

First issue:
The stty characteristics are different with readline (we don't call
YottaDB's setterm()). The main differences are:

- `keypad_local` is emitted after text, not before
- The terminal sends a different sequence of crlf for new lines

I modified the test so that it's still valid but can work on both
readline terminal set-up and YottaDB terminal set-up.

Second issue:
Interrupts that come too quickly cause readline to start losing
characters that are typed. This is not a problem on one of the faster
machines, but shows up on slower machines. For now, we work around
this by reducing interrupt frequency to be between 0-0.9 sec, rather
than the original 0-0.09 sec. See
https://gitlab.com/YottaDB/DB/YDB/-/merge_requests/1382#note_1632232514
for reasoning.

4. Fix r132/ydb612, r132/ydb712, v55000/GTM6940, r140/ydb1029 test
   output for readline

There are several separate issues:

- How we handle interrupts in YottaDB readline code is before issuing
  the prompt (actually we can do it after, but I would rather not
  complicate the code for pre-existing behavior); so the ref files are
  updated so that the output of "YDB>" will show up in the correct spot.
  Uses new AWK defs YDB_READLINE and NON_YDB_READLINE.
- r132/ydb712 has expect race conditions. I fiddled with the script
  until it can pass in both readline and non-readline cases.
- r132/ydb712 When interrupts are received, echo is turned on/off for
  readline. If we don't wait till echo is turned off, the output of
  "halt" show ups twice, once due to our typing (echo), and once after
  the typed text is processed by readline. Wait for echo off first, so
  that "halt" will only show up once.
- r132/ydb612 has the same echo on/off issue. This time, it is with
  CTRL-A being echoed to the screen as "^A" rather than processed as a
  control sequence.

5. Fix r124/ydb350 ref file to display the different stty settings for
   readline and different strace output for readline on certain Linux
   Kernels (extra seds for TCSETS/SNDCTL_TMR_STOP or TCSETSW).

6. Disable expect/gtm7994 for ydb_readline=1. See comments for why that
   was done.

7. Fix ref file for r132/ydb612 for ydb_readline=1 (different output for
   different versions of Readline)

8. Disable v62002/gtm3912 for ydb_readline=1. See comments for why that
   was done.

9. v63000/gtm6388 test counts for strace writes increased, when they
   shouldn't have. It turns out that the readline calls to write to the
   history file used the same file descriptor as the MUPIP calls. This
   is fixed by creating a new AWK file that counts the strace writes
   between when MUPIP opens a file and when it closes it. The problem is
   now fixed.

10. Tests r132/ydb704 and v52000/D9G12002636 are disabled for readline
    v7. For r132/ydb704, readline v7 has a memory issue described in
    this mail message (https://lists.gnu.org/archive/html/bug-readline/2017-03/msg00010.html),
    for v52000/D9G12002636, readline sends new lines when invoked again
    in readline v7 after an interrupt--not an issue in v8.