[YDB#835] [V70001] [GTM-7569] Fix v62000_0/gtm7569 test to increase buffers to deal with V70001s increased BT levels
This test started failing as soon as V70001 was merged. It is supposed to get a MAXBTLEVEL error but started getting
a TRANS2BIG error. After a lot of investigation, it was explained (here:
https://gitlab.com/YottaDB/DB/YDBTest/-/merge_requests/1682#note_1579035023) that V70001 had a change that increased the
maximum BT level for all databases (including V6 DBs) from 7 to 11 meaning it take a whole lot more buffers to hit the
MAXBTLEVEL error the test expects. The global buffers were increased from 4542 to 34396 to compensate for this V70001
change. This buffer count change was done by changing the expression in the below diff:
```
index db6683f1..8bfcc48a 100644
--- a/v62000/u_inref/gtm7569.csh
+++ b/v62000/u_inref/gtm7569.csh
@@ -17,7 +17,7 @@ source $gtm_tst/com/set_limits.csh
 @ block_size = $MIN_BLOCK_SIZE
 @ key_size = 472
 @ record_size = $MAX_RECORD_SIZE
-@ global_buffer_count = 2 * ($record_size + $key_size) / ($block_size - 50) + 1
+@ global_buffer_count = 2 * ($record_size + $key_size) / ($block_size - 451) + 1
 @ allocation = ((3 * $record_size) / 2) / $block_size
 # Create the database with the above configuration
 $gtm_tst/com/dbcreate.csh mumps 1 $key_size $record_size $block_size $allocation $global_buffer_count >&! db_create.out
```
Note if the 451 value is changed to 450, the test fails if we are running V6 db mode and using BG mode access.