[#522] Fix comments that got mangled due to multiple commits in the MR
I am not sure of the origin of this problem but know that I've changed these same
files in the same way 3 separate times yet these unwanted changes keep re-emerging.

I noticed this only after #522/!1682 was committed so this is a cleanup merge. This
is only changing comments so no functional change.