[YDB#835] [#536] [V7.0-001] [GTM-8010] Add v70001/gtm8010 test to verify the EXCEPTION value used correctly on open of /dev/null
GT.M versions prior to V70001 and YottaDB versions prior to V70001 being merged into it (r1.38 and earlier) all
fail this test. Note though that when this test fails with r1.38, the debug version assert fails (in stp_gcol_src.h
line 839) while the pro version fails like GT.M release do (with STACKCRIT errors). As noted in the driving script,
this failure does not produce a SIG-11 like the release note mentions but what we see with the STACKCRIT errors and
even the assert failure in stp_gcol_src is the exact same failure - a negative length is calculated for the exception
string if the string is greater than 128 bytes (up to the max of 255). In our case, this causes an odd loop as it
drives an INDRMAXLEN error that causes us to go through mum_tstart, dm_start and loop back around and get the same
error chewing up a small bit of the stack for each error until we get the STACKCRIT which our $ETRAP handler deals
with and unwinds it all. In the assert failure case, again, the string length is negative which causes some asserts
to trip. Not sure why upstream did not see these additional failure methods or just didn't bother to note them down.

This test starts with around a 32 char exception string, increasing it by 8 or 9 each of 26 iterations. This takes us to
our maximum exception string of 255 bytes (after adding '000' to the last handler). The handlers don't do anything
except record errors if they get driven. The rest of the handler is just a series of generated SET commands to take
up space. Each iteration opens the /dev/null device, writes a small string and closes it. Errors at each iterations
are recorded.