[#364] Fix YDBTest Docker Set-up to run again
Since the previous MR for #364 was merged, it was never integrated into
a pipeline, and never used internally, and as a result, when code
changed things stopped working. We have the intention of getting this
into the YDB pipeline, and having an external contributor use this to
author tests, so this MR brings it up to date again.

Here are the things that needed to be fixed:

- Bring system up to latest version of Debian
- Add missing packages python3, ss, netstat, Java JDK.
- Add /sbin/ to PATH for ldconfig (issue with latest debian).
- GPG set-up (previously not required in the test sytem; added by commit
  d5961ee73eb77bedc6d6eeea37f5bbf66831fdb0): This was complicated as
  the set-up scripts are not part of the test system in the first place, and
  also required a tty. So I did the following:
  - New script `docker/setup_gpg.csh` to do it.
  - Modify on-the-fly `gen_keypair.sh` from YDBEncrypt repo to remove
    stty usages.
- `setenv: Too many arguments`: Error from com/do_random_settings.csh:
  Avoid by setting gtm_test_v6_dbcreate_rand_ver to current version and
  setting gtm_test_use_V6_DBs to 0.