[YDB#835] [#534] [V7.0-001] [GTM-4814] Add v70001/gtm4814 to verify M-profiling (VIEW "TRACE") restored after ZSTEP
This test tests interference between M Profiling and ZSTEP operation. First, the test starts a trace
then executes a few zstep statements in the middle of the trace. The result (with V7.0-001 merged) is
that the statements that are ZSTEPed over do not appear in the trace (which was expected). And once the
test was resumed, the ZSTEPs likewise resumed. Previously, the trace just stopped at the first zstep
never to restart again.

The second test turns that around and tries run some zsteps, then starts tracing to see if it found
any interference as the release note indicated. This test confirms that some interference is seen.