[#529] [#501] Fix manually_start/gds_max_blk test with envvar in output comments causing issues
One of the comments echoed from the test to the reference file contained a usage of the
$ydb_test_4g_db_blks envvar. The reference was not escaped either in the code (echo) that
generated the comment nor in the reference file. In addition the "echo" commands used by
the code output the comment using double quote (") characters to delimit the comment literal
so the envvar had substitution done. So unless the value at test time was 0, the output
comment did not equal the comment in the reference file. This was what that looked like:
```
--- gds_max_blk/gds_max_blk.diff ---
117c117
< # This is because we use 0 to create a 'hole' in the database that contains nothing so
---
> # This is because we use 19122925 to create a 'hole' in the database that contains nothing so
```
What was SUPPOSED to happen was the variable name $ydb_test_4g_db_blks was supposed to appear
in both the output and reference files. Changing the literal text delimiter on the "echo"
commands from double quote to single quote fixed the problem.