[YDB#835] [#535] [V7.0-001] [GTM-5148] Modify v54003/maxregseqno test to verify new REPLAHEAD message with additional information messages
This test modifies the v54003/maxregseqno test - whose primary purpose is to setup the situation
and force a REPLAHEAD (previously a confusing REPL_ROLLBACK_FIRST message). But the test currently
only verifies that a REPLAHEAD message appeared in the log without verifying the additional
information messages. This change to the test verifies that the additional information described
in the release note is present.