[YDB#835] [#539] [V7.0-001] [GTM-9057] Add test v70001/gtm9057 to verify MUPIP JOURNAL -EXTRACT to a FIFO device works correctly
Test fires up a 'reader' process which creates a FIFO with the name `fifo.input` and
reads from it until a special string 'doneDoneDONE' is read at which time it shuts down.

When the reader process reads the journal records fed to it by `MUPIP JOURNAL -EXTRACT`,
it processes them to eliminate variable fields. The 2nd, 3rd, and 4th fields of all journal records
(except the first header record), contain the date/time, transaction number, and process-id
of the process which will vary every time the test is run. In addition, when the record
type (first field delimited by '\') is '01', then fifth and sixth fields also are varying
fields containing the system-name and userid of the process respectively.

Each variable data field is replaced with the string '##FILTERED##' which is a value that
is used to eliminate variable data throughout the test system.

The reason for this is the journal extract is copied to the reference file so we know when
or if it changes.