[YDB#835] [#533] [V7.0-001] [GTM-4272] Fix r124 test failure on RHEL systems due to com/set_gtm_machtype.csh regression in 81a7f6d8
Our Redhat 7 and 8 systems both failed in r124/instream.csh where it called com/set_gtm_machtype.csh.
The problem was in determining whether the copy_file_range() entry point was available or not
(kernel version >= 5.3 and glibc version > 2.27). When the version being considered is less-than 5.3,
the sort command used returned a non-zero return code and on these RedHat systems, that would cause
the script to abort. The script was changed so we no longer generate any non-zero return codes when
making this determination. My thanks to @nars1 for suggesting the fix.

This issue is discussed with more context here: https://gitlab.com/YottaDB/DB/YDBTest/-/merge_requests/1747#note_1668728283