[YDB#835] [V70001] [GTM-4272] Disable second stage of v70001/gtm4272 on ARM systems
The second stage of this tests run a backup with the -SHOWPROGRESS option but on the ARM
systems (both AARCH64 and ARMVXL), the command produces no BKUPPROGRESS messages as it
does on x8664. So this second stage is disabled on ARM.

The reason we are disabling is we don't think it is worth spending time investigating
why BKUPNOTPROGRESS does not show up and since the code that generates it should be
Linux portable, it is okay to test it just on the x86_64.