[YDB#835] [V70001] [#542] [GTM-9324] New sudo/gtm9324 subtest to verify ZSTEP restored after MUPIP INTRPT/ZTIMEOUT and ZSTEP subject to ZBREAK restriction
Test has two parts:
- Part 1 tests that ZSTEP resumes after being interrupted by either $ZINTERUPT or $ZTIMEOUT.
- Part 2 tests that both ZSTEP and ZBREAK respond to the same restrict.txt file restrictions
along with an additional test to verify that when BREAK is restricted, that the ZSTEP OUTOF
command in a debug build can proceed without ASSERT failing.

First part is tested by using ^sstep to initiate a ZSTEP trace, interrupting it by both of
the above mentioned methods and verifying in the trace that the interrupt was seen and that
ZSTEP tracing resumed after the interrupt. This part definitely fails on previous releases.

The second part makes sure both ZSTEP and ZBREAK are subject to the restrictions placed on
ZSTEP. The test uses a copy of the release being tested so we can create a restriction file
in $gtm_dist to verify that both commands become prohibited if a restriction is entered for
ZSTEP.