[YDB#835] [V70001] [YDBTest#548] [GTM-9388] Verify ZSHOW "B" when ZBREAK used with no or null action argument displays correctly
As noted in the issue, this test is only a small piece of the issue. The bulk of the testing for the
issue was satisfied in several tests that were modified for the v70001 merge that include break points
with non-null code arguments. All this test needs to test is the missing and null code arguments of
a ZBREAK command.