[YDB#835] [V7.0-001] [#560] [GTM-9429] New v70001/gtm9429 subtest verifies $QLENGTH() and $QSUBSCRIPT() checking for canonic references
Created 5 cases that work for both functions for a total of 10. Test uses a table containing the list of
testcases so it is easy to add them.

The test cases are placed in a comment array at the end and are fetched using $TEXT(). There are actually 2
comment tables - one for nonUTF8 and one for UTF8 cases. The nonUTF8 cases are always run but the UTF8
cases are only run when UTF8 mode is randomly (or overtly) selected.