[#364] Fix pipeline failure due to race condition
This pipeline failed with exit code 99:
https://gitlab.com/nars1/YDB/-/jobs/5804680318

The exit code 99 comes from line 32 of `docker/run.csh`.

```
15 # needed to create the /var/log/messages or /var/log/syslog file; as otherwise it won't exist at start
16 logger test
17
18 # Fix length of columns to prevent wrapping issues
19 #setenv COLUMNS 4096
20 # Debug
21 #stty -a
22
23 # Next two seds, fix the serverconf.txt file
24 ## Correct the host; as it differs each time we start docker
25 sed -i "s/HOST/$HOST/" /usr/library/gtm_test/serverconf.txt
26 ## Correct the syslog file too
27 if ( -f /var/log/messages ) then
28         sed -i 's|LOG|/var/log/messages|' /usr/library/gtm_test/serverconf.txt
29 else if (-f /var/log/syslog ) then
30         sed -i 's|LOG|/var/log/syslog|' /usr/library/gtm_test/serverconf.txt
31 else
32   exit 99
33 endif
```

My guess of what happened is that the command in line 16 to create the log file
does it asynchronously, and by the time we get to 27 and 29, it doesn't exist
yet, and thus we hit 32.

Since we are running on Debian, we solve this by waiting for the file
`/var/log/syslog` to exist prior to continuing.