[#718] Run changed tests on pipeline
The purpose of this commit is to guess which tests need to be run based
on a `git diff` and then run these tests in the pipeline. The tests use
the docker image `registry.gitlab.com/yottadb/db/ydbtest` as the basis
of running the tests; that docker image is generated in the post-merge
pipeline in YDBTest after each commit.

The exact workflow is in `docker/pipeline-test.csh`; I hope the comments
make it understandable.

Here's the overall algorithm:

1. If `instream.csh` or a file in `inref` changes, run the whole test
   suite in replic and non-replic mode. These changes can be combined
   into a single run as the test system will let us do `-t xxx -t yyy`.
2. If a file inside `u_inref` or `outref` changes, run only that test,
   and don't bunch up the tests together as the test system is incapable
   of handling multiple test/subtest pairs. We check if the test is a
   replic or a non-replic test, and run it appropriately. If the test
   was previously run as part of step 1, it won't run again.

Some notes on some unusual/difficult issues we needed to handle:

1. The test system emphatically wants host names to not contain dashes,
   as it converts the hostname into C Shell variables in various places
   (e.g. `com/gtm_test_ipv6_random.csh`) resulting in hard to trace
   `setenv: too many arguments` errors. There is nothing this commit
   does per se to remedy this; instead the Gitlab pipeline `config.toml`
   file has been now configured so that it runs on static hosts with no
   dashes in their names.
2. We use `rsync` in a script to update the test system; it's not
   installed by default. This commit adds it both in the Dockerfile that
   generates `registry.gitlab.com/yottadb/db/ydbtest`, AND in the
   `.gitlab-ci.yml` file; but it can be removed from the latter after
   merging this commit.

The pipelines takes 3-4 minutes to run the first pipeline of the day
(mainly to pull the docker image) and then each run after that takes
about a minute.

The test systems output will be available as artifacts.

To test this, you need to create a commit on top of this branch, and
push it so that the pipeline will run it.

Sample pipelines:

- https://gitlab.com/shabiel/YDBTest/-/jobs/7879435850: Change
  `r130/inref/ydb484.m`: This results of the run of the `r130` suite
  (replic & non-replic).
  This suite is one of the few that does not run on Docker, mainly due
  to the lack of availability of `gtm_curpro`. You can check the
  artifcats due to failure.
- https://gitlab.com/shabiel/YDBTest/-/jobs/7879562541: Change
  `r136/outref/ydb869.txt v63007/instream.csh v63007/outref/gtm7318.txt
  v63007/u_inref/gtm7318.csh v70005/u_inref/audit_logging-gtmf171004.csh`:
  This results in runs of r136/ydb869 (non-replic),
  v70005/audit_logging-gtmf171004 (non-replic), v60007 (both replic and
  non-replic).

After this is merged, there are a few TODOs:

- Make the pipeline `config.toml` changes permanent (commit to
  YDBDevOps) (currently at
  https://gitlab.com/YottaDB/Util/YDBDevOps/-/merge_requests/183)
- Remove rsync install from `.gitlab-ci.yml`
- Add our ARM64 runner so that it runs tests as well