[#538] [GTM-8843] Concurrency fix for subtests v70001/gtm8843_26 to gtm8843_29
Root cause: The client and server are separate processes, and they wait for each other at several checkpoints, e.g. the client sends a request only when server is already up and running. At the last checkpoint, one does the test itself and the other waits for it. When they both pass the checkpoint, they both print a "finished" message and exit. As they are separate processes, sometimes the foreground process will exit long before the background process, and the shell script will kill it as a remnant process, before it can print the "finished" message.

Fix: background process acquires a lock at the startup, foreground process waits it at exit, silently (it's important, because to not to repeat the same issue).