[#585] New v70002/indirection-gtmde201393 subtest verifies indirection ignores comments and correctly handles empty string arguments
Test features added in GT.M V7.0-002 [GTM-DE201393], specifically:
- that indirection of the form @x@y ignores any comment at the end of the value of x; previously such a comment stopped the evaluation so the subscripts after the second @ were ignored.
- if the first (x) expression evaluates to the empty string, that GT.M produces a VAREXPECTED error; previously that scenario produced incorrect results, sub-optimal errors or, in the worst case damage to an item at the top of the heap.
- also, if the second (y) expression is the empty string, it no longer has the potential to lead to incorrect results or sub-optimal error reports.

Berwyn and Narayanan have been unable to think of an indirection example that could exercise the bullet-3 code path, nor do any existing tests exercise this code (see this thread). It is clear that it is an unusual fringe case, and so it will not be tested.