[GTM-F135428] New v70003/gtmf135428 subtest for test $ZSOCKET() returns an empty string when given an out of range index
Release note (from http://tinco.pair.com/bhaskar/gtm/doc/articles/GTM_V7.0-003_Release_Notes.html#GTM-F135428)

> If the index for the $ZSOCKET() function is outside the range
> of attached sockets for the specified SOCKET device, $ZSOCKET()
> returns an empty string. Previously, if prior actions on the
> SOCKET device removed one or more sockets, $ZSOCKET() could
> return stale or invalid information, or cause a segmentation
> violation (SIG-11).

The test is a server-client pair. The steps seems crazy, but
all of them are needed to trigger a SIGSEGV on previous
versions.

The client connects to the server, and sends a character 'c'.
When the server receives the character, it closes the client
connection.

Then, the server closes the connection again, and catches the
SOCKNOTFND error.

After the server tried to close the already closed connection,
the client attempts to sending an 'x' character.

The server waits (with LOCK mechanism) the client to send the
message, then lists the socket connections, retrieving
information with $ZSOCKET command. It should report empty
string for the closed client connection.