Fix rare sync issue in v70001/gtm8843_16 subtest
Rare cases, v70001/gtm8843_16 subtest failed,
because the opposite side is sending data too quickly,
the actual side is missing the incoming data.

```
< 6    server: non-blocking "READ" query: READ
---
> 6    server: non-blocking "READ" query:
```

The fix is trivial: instead of waiting (HANG) some, added
timeout to WRITE /WAIT instruction.

Hopefully, no test machine will be so slow that sender side
sends data only after receiving side timeouts.
It was almost always working with hang=0.5, now timeout=2s.

(If so, adding some locking might help, keeping the threads
better synchronized, but it's also not a 100% guarantee to
not to miss data when the machine is extremly slow.)