Fix two old HPUX tests that use `which` in a way that stops the test.
Issue: If `which` returns false then the test will quit early since tests
are run with `tcsh -e`.

This test is only run on obsolete HPUX Linux, so this fix will never be run.
However, it could be copied into new code since it is an example of using `which`
to check for the existence of a command, which would cause bugs in future tests.
So best to fix it.